+++ src/components/admin/withdrawals/CreateForm.tsx
@@ -18,10 +18,10 @@ import { useCampaignList } from 'common/hooks/campaigns'
 import { useVaultsList } from 'common/hooks/vaults'
 import GenericForm from 'components/common/form/GenericForm'
 import FormTextField from 'components/common/form/FormTextField'
-import BankAccountSelect from 'components/bankaccounts/BankAccountSelect'
+import BankAccountSelect from 'components/admin/bankaccounts/BankAccountSelect'
 import CampaignSelect, { SetFieldValueType } from 'components/campaigns/CampaignSelect'
-import VaultSelect from 'components/vaults/VaultSelect'
-import PersonSelect from 'components/person/PersonSelect'
+import VaultSelect from 'components/admin/vaults/VaultSelect'
+import PersonSelect from 'components/admin/person/PersonSelect'
 import { Currency } from 'gql/currency'
 import { fromMoney, toMoney } from 'common/util/money'
 import { v4 as uuidv4 } from 'uuid'
