+++ src/components/admin/withdrawals/EditForm.tsx
@@ -24,11 +24,11 @@ import { endpoints } from 'service/apiEndpoints'
 import { useCampaignList } from 'common/hooks/campaigns'
 import FormTextField from 'components/common/form/FormTextField'
 import GenericForm from 'components/common/form/GenericForm'
-import CurrencySelect from 'components/currency/CurrencySelect'
+import CurrencySelect from 'components/common/currency/CurrencySelect'
 import CampaignSelect from 'components/campaigns/CampaignSelect'
-import VaultSelect from 'components/vaults/VaultSelect'
-import PersonSelect from 'components/person/PersonSelect'
-import BankAccountSelect from 'components/bankaccounts/BankAccountSelect'
+import VaultSelect from 'components/admin/vaults/VaultSelect'
+import PersonSelect from 'components/admin/person/PersonSelect'
+import BankAccountSelect from 'components/admin/bankaccounts/BankAccountSelect'
 import { Currency } from 'gql/currency'
 import { fromMoney, toMoney } from 'common/util/money'
 import { useVaultsList } from 'common/hooks/vaults'
